requirements.txt
streamlit==1.53.1
pandas==2.2.3
plotly==6.0.0
supabase==2.28.0
python-dotenv==1.2.1
